fix(api): company req validation

+++ internal/application/api.go
@@ -8,7 +8,7 @@ import (
 )
 
 type API interface {
-	GetCompany(id int) (*domain.Company, error)
+	GetCompany(id int64) (*domain.Company, error)
 	GetCompanyUpdates(from, to time.Time, status string) []domain.Company
 }
